requirements.txt
flask>=3.0.0
flask-swagger-ui>=4.11.1
mysql-connector-python>=9.0.0
